+++ app/components/asteroid-list/AsteroidList.module.css
@@ -1,0 +1,36 @@
+.list__head {
+  max-width: 400px;
+}
+
+.list__title {
+  color: var(--White-Base, #FFF);
+  font-family: var(--font-helvetica);
+  font-size: 28px;
+  font-style: normal;
+  font-weight: 600;
+  line-height: 36px;
+  /* 128.571% */
+}
+
+.list__toggler {
+  font-family: var(--font-helvetica), 'Courier New';
+  font-size: 14px;
+  font-style: normal;
+  line-height: 24px;
+}
+
+.list__toggler__button {
+  color: var(--White-Base, #FFF);
+  font-size: 16px;
+  font-weight: 300;
+  text-decoration: underline;
+  background-color: rgba(0, 0, 0, 0);
+  border: none;
+  cursor: pointer;
+}
+
+.list__toggler__button:disabled {
+  font-weight: 700;
+  text-decoration: none;
+  cursor: auto
+}
